Fixing cases where method completes after client has given up on it.
When using a timeout in remote_send_with_timeout or wait_for_done, if
the method was in progress we would leave results around in Redis that
never got cleaned up. Now RedisSpace.result_set will check to make
sure the method is still active; if not, results will get thrown away.