When an exception is raised in the method_provider, the backtrace will now also be sent back to the method_sendor. Before, only the exception class and message would get sent back.
The receiving end of the exception will combine its own
stack trace with the exceptions backtrace so that the
backtrace appears as if the method_sender directly called
the method_provider.  This provides the user with a complete
end-to-end trace across LiveResource.

The final backtrace does not have any LiveResource method
traces in it.  Instead, it only has a single 'via LiveResource'
line in it.  Someone debugging one of these exceptions
won't care about LiveResource traces anyway.