Fix a bug that could cause certain exceptions (e.g. NameError) to not get returned through LiveResource properly.  The issue happens when RedisClient#deserialize tries to de-YAML'ize the exception object causing another exception to be raised.
The way exceptions are passed back has been changed so that the actual
object is never sent back.

Old way:
    [exception_object, message, backtrace]
New way:
    ['Exception', exception class name, message, backtrace]

The string 'Exception' is in there to help solidify the fact that it is
an exception we're passing back.  Deserialize will check for 'Exception'
and that the array's length is 4.

test/exception_backtrace_test.rb:
    Update this unit test to specifically test the bug being fixed.

    A NameError is now the type of exception being passed back instead
of 'raise "failure"'.