Add the ability to filter LiveResource log messages.  From Will Andrews...
Allow using environment variables to filter LiveResource logging.

Example:
env LIVERESOURCE_DEBUG_IGNORE="publish:set:get" \
/usr/local/zLine/workers/bin/worker_manager run zfs_worker

...will ignore LiveResource debug logs that begin with "publish",
"set", and "get".  This is extremely useful when debugging a worker
that processes large amounts of data through LiveResource.

Add the following two methods to this module:
- #ignore_log?(level, str): Returns whether the log should be ignored,
  given the desired log level and log string.
- #ignore_log(level): Returns an array of strings representing the
  starting words of logs to ignore.  By default, when this method is
  first invoked, it will generate an array that splits the environment
  variable LIVERESOURCE_<level.upcase>_IGNORE on colons.  These are
  stored as level=>array pairs in @loghelper_ignores.

In each log level method, generate the string and check whether it
should be ignored before passing it on to the logger.