Commit on behalf of Jason Stevens:
Fixed bug in live resource:

merge_backtrace expects to find regexp
/lib\/live_resource\/resource_proxy/ somewhere in the 'sender_trace'
stack trace, but the 'sender_trace' is the results of caller(), which
is really caller(1), which per the Ruby documentation for caller
means that we nix the top-most element of the stack trace, which is
the only element *guaranteed* to match regexp
/lib\/live_resource\/resource_proxy/

Fix is to explicitly call caller(0), which ensures that we have at
least one element that matches /lib\/live_resource\/resource_proxy/
Also, if there are multiple elements that match
/lib\/live_resource\/resource_proxy/, we want to nix them all - not
just one of them.

Verify that "minimal stack" backtrace merges work (stack traces that
only have the resource_proxy as a single element in the trace).