method_result will now also log the result when the caller deleeted the method.
method_result is called when the method call, on the method provider
side, returns.  When the method has been deleted, a warning is logged
indicated that condition.  This change also logs the actual method
result.  This scenario likely happens when the method caller gave up
waiting for the method provider to finish executing the method.