Close a race condition in the ForwardContinueTest#test_message_path test that was causing intermittent test failures.
test/test_helper.rb:
Add a redis_keys helper

test/method_forward_continue_test.rb:
Change the test_message_path test to compre the redis keys
instead of just the redis "dbsize".  This makes the test easier to debug
when it fails since the user will see what keys are different.

test/method_forward_continue_test.rb:
Use a queue so the class3 instance can signal the main test thread when
method_done is called and thus is safe to check the contents of Redis to
verify that it has cleaned everything up.  This is necessary since the
resource instance cleans up its in progress methods after returning the
result of the method.